fix: useCreateCommunity 코드 수정
